refactor: move create relay logic to dedicated provider
