+++ lib/app/features/ion_connect/providers/ion_connect_notifier.c.dart
@@ -5,19 +5,18 @@ import 'dart:convert';
 
 import 'package:ion/app/exceptions/exceptions.dart';
 import 'package:ion/app/extensions/extensions.dart';
-import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
-import 'package:ion/app/features/chat/providers/user_chat_relays_provider.c.dart';
 import 'package:ion/app/features/ion_connect/ion_connect.dart' as ion;
 import 'package:ion/app/features/ion_connect/ion_connect.dart' hide requestEvents;
 import 'package:ion/app/features/ion_connect/model/action_source.dart';
 import 'package:ion/app/features/ion_connect/model/auth_event.c.dart';
 import 'package:ion/app/features/ion_connect/model/event_serializable.dart';
 import 'package:ion/app/features/ion_connect/model/ion_connect_entity.dart';
+import 'package:ion/app/features/ion_connect/providers/active_relays_provider.c.dart';
 import 'package:ion/app/features/ion_connect/providers/auth_challenge_provider.c.dart';
 import 'package:ion/app/features/ion_connect/providers/ion_connect_cache.c.dart';
 import 'package:ion/app/features/ion_connect/providers/ion_connect_event_parser.c.dart';
 import 'package:ion/app/features/ion_connect/providers/ion_connect_event_signer_provider.c.dart';
-import 'package:ion/app/features/ion_connect/providers/active_relays_provider.c.dart';
+import 'package:ion/app/features/ion_connect/providers/relay_creation_provider.c.dart';
 import 'package:ion/app/features/ion_connect/providers/relays_provider.c.dart';
 import 'package:ion/app/features/user/model/user_chat_relays.c.dart';
 import 'package:ion/app/features/user/model/user_relays.c.dart';
@@ -45,7 +44,9 @@ class IonConnectNotifier extends _$IonConnectNotifier {
 
     return withRetry(
       ({error}) async {
-        relay = relay ?? await _getRelay(actionSource, dislikedUrls: dislikedRelaysUrls);
+        relay ??= await ref
+            .read(relayCreationProvider.notifier)
+            .getRelay(actionSource, dislikedUrls: dislikedRelaysUrls);
 
         if (_isAuthRequired(error)) {
           await sendAuthEvent(relay!);
@@ -148,7 +149,9 @@ class IonConnectNotifier extends _$IonConnectNotifier {
 
     yield* withRetryStream(
       ({error}) async* {
-        relay = await _getRelay(actionSource, dislikedUrls: dislikedRelaysUrls);
+        relay = await ref
+            .read(relayCreationProvider.notifier)
+            .getRelay(actionSource, dislikedUrls: dislikedRelaysUrls);
 
         if (_isAuthRequired(error)) {
           try {
@@ -241,161 +244,6 @@ class IonConnectNotifier extends _$IonConnectNotifier {
     return isSubscriptionAuthRequired || isSendEventAuthRequired;
   }
 
-  Future<IonConnectRelay> _getRelay(
-    ActionSource actionSource, {
-    Set<String> dislikedUrls = const {},
-  }) async {
-    switch (actionSource) {
-      case ActionSourceCurrentUser():
-        final pubkey = await ref.read(currentPubkeySelectorProvider.future);
-        if (pubkey == null) {
-          throw UserMasterPubkeyNotFoundException();
-        }
-
-        final userRelays = await _getUserRelays(pubkey).then((userRelays) => userRelays.data.list);
-        final relays = _userRelaysAvoidingDislikedUrls(userRelays, dislikedUrls)
-            .map((relay) => relay.url)
-            .toList();
-
-        final lastUsedRelays = await _getLastUsedRelay(
-          relays,
-          dislikedUrls,
-          actionSource.anonymous,
-        );
-        if (lastUsedRelays != null) return lastUsedRelays;
-
-        return _getAndActivateRelay(relays.random, actionSource.anonymous);
-
-      case ActionSourceUser():
-        final userRelays =
-            await _getUserRelays(actionSource.pubkey).then((userRelays) => userRelays.data.list);
-
-        final relays = _userRelaysAvoidingDislikedUrls(userRelays, dislikedUrls)
-            .map((relay) => relay.url)
-            .toList();
-
-        final lastUsedRelays = await _getLastUsedRelay(
-          relays,
-          dislikedUrls,
-          actionSource.anonymous,
-        );
-
-        if (lastUsedRelays != null) return lastUsedRelays;
-
-        return _getAndActivateRelay(relays.random, actionSource.anonymous);
-
-      case ActionSourceCurrentUserChat():
-        final pubkey = await ref.read(currentPubkeySelectorProvider.future);
-        if (pubkey == null) {
-          throw UserMasterPubkeyNotFoundException();
-        }
-
-        final userRelays =
-            await _getUserChatRelays(pubkey).then((userChatRelays) => userChatRelays.data.list);
-
-        final relays = _userRelaysAvoidingDislikedUrls(userRelays, dislikedUrls)
-            .map((relay) => relay.url)
-            .toList();
-
-        final lastUsedRelays = await _getLastUsedRelay(
-          relays,
-          dislikedUrls,
-          actionSource.anonymous,
-        );
-        if (lastUsedRelays != null) return lastUsedRelays;
-
-        return _getAndActivateRelay(relays.random, actionSource.anonymous);
-
-      case ActionSourceIndexers():
-        final indexers = await ref.read(currentUserIndexersProvider.future);
-        if (indexers == null) {
-          throw UserIndexersNotFoundException();
-        }
-
-        final relays = _indexersAvoidingDislikedUrls(indexers, dislikedUrls);
-
-        final lastUsedRelays = await _getLastUsedRelay(
-          relays,
-          dislikedUrls,
-          actionSource.anonymous,
-        );
-        if (lastUsedRelays != null) return lastUsedRelays;
-
-        return _getAndActivateRelay(relays.random, actionSource.anonymous);
-
-      case ActionSourceUserChat():
-        final userChatRelays = await _getUserChatRelays(actionSource.pubkey);
-        final relays = _userRelaysAvoidingDislikedUrls(userChatRelays.data.list, dislikedUrls)
-            .map((relay) => relay.url)
-            .toList();
-        {}
-
-        final lastUsedRelays = await _getLastUsedRelay(
-          relays,
-          dislikedUrls,
-          actionSource.anonymous,
-        );
-        if (lastUsedRelays != null) return lastUsedRelays;
-
-        return _getAndActivateRelay(relays.random, actionSource.anonymous);
-
-      case ActionSourceRelayUrl():
-        final relay = actionSource.url;
-
-        final lastUsedRelays = await _getLastUsedRelay(
-          [relay],
-          dislikedUrls,
-          actionSource.anonymous,
-        );
-        if (lastUsedRelays != null) return lastUsedRelays;
-
-        return _getAndActivateRelay(relay, actionSource.anonymous);
-    }
-  }
-
-  Future<IonConnectRelay?> _getLastUsedRelay(
-    List<String> userRelays,
-    Set<String> dislikedUrls,
-    bool anonymous,
-  ) async {
-    final activeRelaysSet = ref.read(activeRelaysProvider);
-    if (activeRelaysSet.isEmpty) return null;
-
-    final availableRelays = userRelays
-        .where(
-          (relayUrl) => activeRelaysSet.contains(relayUrl) && !dislikedUrls.contains(relayUrl),
-        )
-        .toList();
-
-    if (availableRelays.isEmpty) return null;
-
-    return await ref.read(
-      relayProvider(availableRelays.random, anonymous: anonymous).future,
-    );
-  }
-
-  Future<IonConnectRelay> _getAndActivateRelay(String url, bool anonymous) async {
-    final selectedRelay = await ref.read(relayProvider(url, anonymous: anonymous).future);
-    ref.read(activeRelaysProvider.notifier).addRelay(selectedRelay.url);
-    return selectedRelay;
-  }
-
-  Future<UserRelaysEntity> _getUserRelays(String pubkey) async {
-    final userRelays = await ref.read(userRelayProvider(pubkey).future);
-    if (userRelays == null) {
-      throw UserRelaysNotFoundException();
-    }
-    return userRelays;
-  }
-
-  Future<UserChatRelaysEntity> _getUserChatRelays(String pubkey) async {
-    final userRelays = await ref.read(userChatRelaysProvider(pubkey).future);
-    if (userRelays == null) {
-      throw UserChatRelaysNotFoundException();
-    }
-    return userRelays;
-  }
-
   IonConnectEntity _parseAndCache(EventMessage event) {
     final parser = ref.read(eventParserProvider);
     final entity = parser.parse(event);
@@ -403,22 +251,5 @@ class IonConnectNotifier extends _$IonConnectNotifier {
       ref.read(ionConnectCacheProvider.notifier).cache(entity);
     }
     return entity;
-  }
-
-  List<UserRelay> _userRelaysAvoidingDislikedUrls(
-    List<UserRelay> relays,
-    Set<String> dislikedRelaysUrls,
-  ) {
-    final urls = relays.where((relay) => !dislikedRelaysUrls.contains(relay.url)).toList();
-    if (urls.isEmpty) return relays;
-    return urls;
-  }
-
-  List<String> _indexersAvoidingDislikedUrls(List<String> indexers, Set<String> dislikedUrls) {
-    var urls = indexers.where((indexer) => !dislikedUrls.contains(indexer)).toList();
-    if (urls.isEmpty) {
-      urls = indexers;
-    }
-    return urls;
   }
 }
